requirements.txt
flask
gunicorn
flask-cors
requests
google-auth
diskcache
pydub
google-generativeai>=0.5.0